Move some AddCards specific code to NoteCreator.svelte
